+++ src/utils/validations/card/cvcNumber.ts
@@ -1,5 +1,5 @@
-import { MAX_LENGTH } from "../../constants/constants";
-import { hasExactLength, isPositiveInteger } from "./common";
+import { MAX_LENGTH } from "../../../constants/constants";
+import { hasExactLength, isPositiveInteger } from "../common";
 
 export const isErrorCvcNumber = (value: string) => {
   if (!isPositiveInteger(value)) return "숫자만 입력 가능합니다";
